requirements.txt
numpy
pandas
scipy
scikit-learn
lightgbm
pyyaml
matplotlib
networkx
